Create Project1
I created this script to help me plan out my weekly budget. As per the description on this assignment I created a script that is a start to a larger project. I would like to create a more complex project that can help me budget for an entire month with more ocmplex saving stratigies and extended savings accounts. You can run this script in python and to use it you will simply input the data it asks. Your weekly allowance is is determined by the paramater you input but will automatically give you a $0 allowance if your bills that are due exceed the amount you made that week and subtract the remaining amount from your Current account balence. Also, if your will over draw your account it will give you a warning notice